# docs/swagger.yaml
basePath: /
...
host: 192.168.55.42:8080
info:
  contact: {}
  description: API for an e-commerce platform with OTP-based authentication and product/market
    management.
  title: E-commerce API
  version: "1.0"
paths:
...
  /api/superadmin/market-messages:
    get:
      description: Retrieves all messages sent by markets to superadmin. Requires
        superadmin JWT authentication.
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get all market messages
      tags:
      - Market Messages
...
  /api/superadmin/user-messages:
    get:
      description: Retrieves all messages sent by users to superadmin. Requires superadmin
        JWT authentication.
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get all user messages
      tags:
      - User Messages
...
  /categories:
    get:
      description: Retrieves a paginated list of categories with optional name search.
        Requires superadmin JWT authentication.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Search by category name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get categories
      tags:
      - Categories
...
  /markets:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all markets. Requires JWT authentication.
      parameters:
      - description: 'Duration day for new (default: 7)'
        in: query
        name: duration
        type: integer
      - description: 'Filter new markets (default: false)'
        in: query
        name: is_new
        type: boolean
      - description: 'Filter vip markets (default: false)'
        in: query
        name: is_vip
        type: boolean
      produces:
      - application/json
      responses: {}
      summary: Get all markets
      tags:
      - Markets
  /markets/{id}:
    get:
      description: Retrieves market details and its products by market ID with pagination.
      parameters:
      - description: Market ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Number of products per page (default: 20, min: 1, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Page number (default: 1, min: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Get market by ID
      tags:
      - Markets
  /products:
    get:
      description: Retrieves paginated products with optional category, price range,
        discount, new status, sorting, name search, or random selection for homepage
      parameters:
      - description: Category ID
        in: query
        name: category_id
        type: string
      - description: 'Duration day for new (default: 7)'
        in: query
        name: duration
        type: integer
      - description: 'Page number (default: 1, ignored if random=true)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, ignored if random=true)'
        in: query
        name: limit
        type: integer
      - description: Search by product name
        in: query
        name: search
        type: string
      - description: 'Return 10 random products (default: false)'
        in: query
        name: random
        type: boolean
      - description: Minimum final price
        in: query
        name: start_price
        type: number
      - description: Maximum final price
        in: query
        name: end_price
        type: number
      - description: 'Sort order: cheap_to_expensive, expensive_to_cheap (default:
          by ID)'
        in: query
        name: sort
        type: string
      - description: 'Filter products with discount (default: false)'
        in: query
        name: has_discount
        type: boolean
      - description: 'Filter new products (default: false)'
        in: query
        name: is_new
        type: boolean
      produces:
      - application/json
      responses: {}
      summary: Get all products
      tags:
      - Products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a single product by its ID. Requires JWT authentication.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get a product by ID
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication (e.g., "Bearer <token>")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"